# docs/swagger.yaml
basePath: /
definitions:
...
  createGatewayRouteRequest:
    properties:
      callback_url:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      provider:
        type: string
    type: object
...
  errorResponse:
    properties:
      error:
        type: string
    type: object
...
  gatewayRouteResponse:
    properties:
      callback_url:
        type: string
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      inbound_url:
        type: string
      name:
        type: string
      provider:
        type: string
      revoked_at:
        type: string
      secret:
        type: string
      secret_prefix:
        type: string
      updated_at:
        type: string
    type: object
...
  paginatedResponse-handler_sandboxResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/sandboxResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
    type: object
...
  sandboxResponse:
    properties:
      agent_id:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      external_id:
        type: string
      id:
        type: string
      last_active_at:
        type: string
      status:
        type: string
    type: object
...
  updateGatewayRouteRequest:
    properties:
      callback_url:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      provider:
        type: string
    type: object
...
host: api.dev.usehivy.com
info:
  contact: {}
  description: Proxy runtime for LLM API credentials.
  title: Hivy API
  version: "1.0"
paths:
...
  /v1/agents/{id}/gateway-routes:
    get:
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gatewayRouteResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List external gateway routes
      tags:
      - agents
    post:
      consumes:
      - application/json
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Gateway route
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/createGatewayRouteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gatewayRouteResponse'
      security:
      - BearerAuth: []
      summary: Create an external gateway route
      tags:
      - agents
  /v1/agents/{id}/gateway-routes/{routeID}:
    delete:
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Gateway route ID
        in: path
        name: routeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Revoke an external gateway route
      tags:
      - agents
    get:
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Gateway route ID
        in: path
        name: routeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gatewayRouteResponse'
      security:
      - BearerAuth: []
      summary: Get an external gateway route
      tags:
      - agents
    patch:
      consumes:
      - application/json
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Gateway route ID
        in: path
        name: routeID
        required: true
        type: string
      - description: Gateway route updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/updateGatewayRouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gatewayRouteResponse'
      security:
      - BearerAuth: []
      summary: Update an external gateway route
      tags:
      - agents
  /v1/agents/{id}/gateway-routes/{routeID}/rotate-secret:
    post:
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Gateway route ID
        in: path
        name: routeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gatewayRouteResponse'
      security:
      - BearerAuth: []
      summary: Rotate an external gateway route secret
      tags:
      - agents
...
  /v1/sandboxes:
    get:
      description: Returns sandboxes for the current organization.
      parameters:
      - description: Filter by status (running, stopped, error)
        in: query
        name: status
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paginatedResponse-handler_sandboxResponse'
      security:
      - BearerAuth: []
      summary: List sandboxes
      tags:
      - sandboxes
  /v1/sandboxes/{id}:
    delete:
      description: Deletes a sandbox from the provider and removes the DB record.
      parameters:
      - description: Sandbox ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a sandbox
      tags:
      - sandboxes
    get:
      description: Returns sandbox details by ID.
      parameters:
      - description: Sandbox ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sandboxResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Get a sandbox
      tags:
      - sandboxes
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Bearer token (JWT or API key).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"